Complete Ability Refactor + POC Custom Ability Event
