+++ src/main/java/me/zenox/superitems/abilities/EmberShootSmall.java
@@ -1,4 +1,4 @@
-package me.zenox.superitems.item.abilities;
+package me.zenox.superitems.abilities;
 
 import me.zenox.superitems.SuperItems;
 import org.bukkit.ChatColor;
@@ -6,9 +6,12 @@ import org.bukkit.Location;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Fireball;
 import org.bukkit.entity.Player;
+import org.bukkit.event.Event;
 import org.bukkit.event.player.PlayerEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.metadata.FixedMetadataValue;
+
+import java.util.function.Consumer;
 
 public class EmberShootSmall extends ItemAbility {
 
@@ -20,7 +23,7 @@ public class EmberShootSmall extends ItemAbility {
     }
 
     @Override
-    public void runExecutable(PlayerEvent event) {
+    public void runExecutable(Event event) {
         PlayerInteractEvent e = ((PlayerInteractEvent) event);
         Player p = e.getPlayer();
