+++ src/main/java/me/zenox/superitems/events/PlayerUseItemEvent.java
@@ -1,10 +1,16 @@
 package me.zenox.superitems.events;
 
 import me.zenox.superitems.SuperItems;
+import me.zenox.superitems.abilities.AttackAbility;
+import me.zenox.superitems.abilities.ItemAbility;
 import me.zenox.superitems.item.ComplexItem;
+import me.zenox.superitems.item.ComplexItemStack;
 import me.zenox.superitems.item.ItemRegistry;
-import me.zenox.superitems.item.abilities.Ability;
+import me.zenox.superitems.abilities.Ability;
+import me.zenox.superitems.util.Util;
+import org.bukkit.entity.Player;
 import org.bukkit.event.*;
+import org.bukkit.event.entity.EntityDamageByEntityEvent;
 import org.bukkit.event.player.PlayerEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.inventory.ItemStack;
@@ -23,34 +29,33 @@ public class PlayerUseItemEvent implements Listener {
 
     @EventHandler
     public void useEvent(Event event) {
-        if (!(event instanceof PlayerEvent)){
-            return;
+        if(event instanceof EntityDamageByEntityEvent){
+            EntityDamageByEntityEvent e = ((EntityDamageByEntityEvent) event);
+            if (e.getDamager() instanceof Player) {
+                Player p = ((Player) e.getDamager());
+                for (Ability ability :
+                        Ability.Slot.uniqueEquipped(p)) {
+                    Util.logToConsole("Ability: " + ability);
+                    if (ability instanceof AttackAbility) {
+                        ability.useAbility(event);
+                    }
+                }
+            }
         }
 
         if(event instanceof PlayerInteractEvent){
             interact(((PlayerInteractEvent) event));
             return;
         }
-
-        PlayerEvent e = (PlayerEvent) event;
-        for (ItemStack item : e.getPlayer().getInventory().getContents()) {
-            if (item == null) continue;
-            ComplexItem complexItem = ItemRegistry.getBasicItemFromItemStack(item);
-            if (complexItem == null) continue;
-            for (Ability ability : complexItem.getAbilities()) {
-                if (!ability.getSlot().evaluate(e.getPlayer().getInventory(), ability)) continue;
-                ability.useAbility(e);
-            }
-        }
     }
 
     private boolean interact(PlayerInteractEvent event){
         ItemStack item = event.getItem();
         if (item == null) return false;
-        ComplexItem complexItem = ItemRegistry.getBasicItemFromItemStack(item);
+        ComplexItemStack complexItem = ComplexItemStack.of(item);
         if (complexItem == null) return false;
         for (Ability ability : complexItem.getAbilities()) {
-            ability.useAbility(event);
+            if (ability instanceof ItemAbility) ability.useAbility(event);
         }
         return false;
     }
