+++ src/main/java/me/zenox/superitems/item/ComplexItem.java
@@ -6,8 +6,8 @@ import com.google.common.primitives.Ints;
 import me.zenox.superitems.SuperItems;
 import me.zenox.superitems.data.TranslatableList;
 import me.zenox.superitems.data.TranslatableText;
-import me.zenox.superitems.item.abilities.Ability;
-import me.zenox.superitems.item.abilities.ItemAbility;
+import me.zenox.superitems.abilities.Ability;
+import me.zenox.superitems.abilities.ItemAbility;
 import me.zenox.superitems.util.Util;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
