+++ src/main/java/me/zenox/superitems/item/superitems/DarkEmberStaff.java
@@ -1,22 +1,16 @@
 package me.zenox.superitems.item.superitems;
 
 import com.archyx.aureliumskills.stats.Stats;
-import me.zenox.superitems.SuperItems;
 import me.zenox.superitems.item.ComplexItem;
-import me.zenox.superitems.item.abilities.EmberAttune;
-import me.zenox.superitems.item.abilities.EmberShoot;
-import org.bukkit.ChatColor;
+import me.zenox.superitems.abilities.EmberAttune;
+import me.zenox.superitems.abilities.EmberShoot;
 import org.bukkit.Material;
-import org.bukkit.NamespacedKey;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.ItemFlag;
 import org.bukkit.inventory.Recipe;
 import org.bukkit.inventory.RecipeChoice;
 import org.bukkit.inventory.ShapedRecipe;
-import org.bukkit.persistence.PersistentDataContainer;
-import org.bukkit.persistence.PersistentDataType;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
