+++ src/main/java/me/zenox/superitems/item/superitems/ObsidianScythe.java
@@ -1,7 +1,6 @@
 package me.zenox.superitems.item.superitems;
 
 import me.zenox.superitems.item.ComplexItem;
-import me.zenox.superitems.item.abilities.ObsidilithScytheAbility;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.inventory.Recipe;
@@ -11,7 +10,7 @@ import java.util.Map;
 
 public class ObsidianScythe extends ComplexItem {
     public ObsidianScythe() {
-        super("obsidian_scythe", Rarity.LEGENDARY, Type.MISC, Material.NETHERITE_HOE, Map.of(), List.of(new ObsidilithScytheAbility()));
+        super("obsidian_scythe", Rarity.LEGENDARY, Type.MISC, Material.NETHERITE_HOE, Map.of(), List.of());
 
         List<String> lore = List.of(ChatColor.RED + "" + ChatColor.ITALIC + "The Densest of All Compact Obisidians Combined to Make this Terror...");
