+++ src/main/java/me/zenox/superitems/item/superitems/VoidScepter.java
@@ -4,8 +4,8 @@ import com.archyx.aureliumskills.stats.Stats;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Multimap;
 import me.zenox.superitems.item.ComplexItem;
-import me.zenox.superitems.item.abilities.VoidWarp;
-import me.zenox.superitems.item.abilities.VoidularRecall;
+import me.zenox.superitems.abilities.VoidWarp;
+import me.zenox.superitems.abilities.VoidularRecall;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.attribute.Attribute;
