+++ src/main/resources/languages/en_US.yml
@@ -78,24 +78,24 @@ item-name-volken_axe: Volken Axe
 ##########################################################################
 
 item-lore-purified_magma_distillate:
-  - '&c&oBurning Hot.'
+  - "&c&oBurning Hot."
   -
-  - '&7Rumors are that a single drop could power a city for generations.'
+  - "&7Rumors are that a single drop could power a city for generations."
 
 item-lore-absolute_ender_pearl:
-  - 'Made from the souls of thousands of endermen.'
+  - "Made from the souls of thousands of endermen."
 
 item-lore-dark_skull:
-  - '&oBurning through your soul.'
+  - "&oBurning through your soul."
 
 item-lore-tormented_soul:
-  - 'A trapped soul, facing torture for generations.'
+  - "A trapped soul, facing torture for generations."
 
 item-lore-titanium_cube:
-  - 'Heavyweight.'
+  - "Heavyweight."
 
 item-lore-magic_toy_stick:
-  - '&c&oMagical.'
+  - "&c&oMagical."
 
 item-lore-soul_crystal:
   - "&8&oThe void-Sphere of Life was an anti-magic device, which could absorb the"
@@ -110,140 +110,155 @@ item-lore-soul_crystal:
   - "&8&oGenerated at: https://app.inferkit.com/demo"
 
 item-lore-fiery_ember_staff:
-  - 'A fiery rod, taken from &cBlaziel &7himself.'
+  - "A fiery rod, taken from &cBlaziel &7himself."
 
 item-lore-dark_ember_staff:
-  - 'A fiery rod, taken from &cBlaziel &7himself.'
+  - "A fiery rod, taken from &cBlaziel &7himself."
   - "Draws power from the user's &bmana."
 
 item-lore-tormented_blade:
-  - 'Did you know that this item used be legendary?'
+  - "Did you know that this item used be legendary?"
 
 item-lore-crucified_amulet:
-  - 'Sacrifices must be made.'
+  - "Sacrifices must be made."
 
 item-lore-desecrator_scale:
-  - 'Tough as nails.'
+  - "Tough as nails."
 
 item-lore-desecrator_claw:
-  - 'Sharper than an blade.'
+  - "Sharper than an blade."
 
 item-lore-desecrator_toe:
-  - 'Stinky.'
+  - "Stinky."
 
 item-lore-crulen_shard:
-  - 'Talk to &e[NPC] &6Elmir&7!'
+  - "Talk to &e[NPC] &6Elmir&7!"
 
 item-lore-pythemion_gem:
-  - 'Really fricking hot.'
+  - "Really fricking hot."
 
 item-lore-psychedelic_orb:
-  - '&oIn order for one to know the world around it, they must know about thyself.'
-  - ''
-  - 'Voodoo magic!'
+  - "&oIn order for one to know the world around it, they must know about thyself."
+  - ""
+  - "Voodoo magic!"
 
 item-lore-ravager_skin:
   - "Rough, isn't it?"
 
 item-lore-tough_fabric:
-  - 'Sewn together with blood, sweat, and tears.'
+  - "Sewn together with blood, sweat, and tears."
 
 item-lore-kevlar:
-  - 'Bulletproof.'
+  - "Bulletproof."
 
 item-lore-totem_pole:
-  - '&eCentralized Power.'
+  - "&eCentralized Power."
 
 item-lore-corrupt_totem_pole:
-  - '&cWith great power, eventually comes &dcorruption.'
+  - "&cWith great power, eventually comes &dcorruption."
 
 item-lore-void_scepter:
-  - 'Forged from the souls of corrupted endermen.'
+  - "Forged from the souls of corrupted endermen."
 
 item-lore-obsidilith_scythe:
   - "Mining away... I don't know.. what to miiine..."
 
 item-lore-dev_stick:
-  - 'A tribute to Andy Zhang. RIP.'
-  - ''
-  - '&8&oTo Andy Zhang: If you ever see this, where the &kF*CK &8&ois my content?'
+  - "A tribute to Andy Zhang. RIP."
+  - ""
+  - "&8&oTo Andy Zhang: If you ever see this, where the &kF*CK &8&ois my content?"
 
 item-lore-volken_axe:
-  - 'A relic of the past. Fragile. &oDo Not Touch.'
-  - ''
-  - 'Did you really think this was &ofragile?'
+  - "A relic of the past. Fragile. &oDo Not Touch."
+  - ""
+  - "Did you really think this was &ofragile?"
 
 ##########################################################################
 ##                              ABILITIES                               ##
 ##########################################################################
 ability-name-magic_missile: Magic Missile
 
-ability-lore-magic_missile:
-  - 'Shoots a magic missile that explodes'
-  - 'on impact and deals massive &cdamage.'
+ability-lore-magic_missile_dev:
+  - "Shoots a magic missile that explodes"
+  - "on impact and deals massive &cdamage."
 
-ability-lore-magic_missile-combust:
-  - 'Shoots a magic missile that explodes'
-  - 'on impact and deals massive &cdamage.'
+ability-lore-magic_missile:
+  - "Shoots a magic missile that explodes"
+  - "on impact and deals massive &cdamage."
   -
-  - '20% chance for the item to &6combust &7and dissapear.'
+  - "20% chance for the item to &6combust &7and disappear."
 
 ability-name-soul_rift: Soul Rift
 
 ability-lore-soul_rift:
-  - 'Deploys a &bsoul crystal&7, which lasts for &a5s &7and'
-  - 'pulls entities in, dealing &ctrue damage &7every second.'
+  - "Deploys a &bsoul crystal&7, which lasts for &a5s &7and"
+  - "pulls entities in, dealing &ctrue damage &7every second."
 
 ability-name-fiery_ember_shoot: Fiery Embers
 
 ability-lore-fiery_ember_shoot:
-  - 'Shoot a &cfireball &7 that explodes'
-  - 'on impact. Fiery Explosions!'
+  - "Shoot a &cfireball &7 that explodes"
+  - "on impact. Fiery Explosions!"
 
 ability-name-dark_ember_attune: Attune
 
 ability-lore-dark_ember_attune:
-  - 'Changes the &battunement &7of this staff,'
-  - 'changing the domain it draws power from.'
-  -
-  - 'Swap between &6Blazeborn &7and &8Darksoul'
+  - "Changes the &battunement &7of this staff,"
+  - "changing the domain it draws power from."
+  - ""
+  - "Swap between &6Blazeborn &7and &8Darksoul"
 
 ability-name-dark_ember_shoot: Dark Embers
 
 ability-lore-dark_ember_shoot:
-  - 'Shoot a projectile based on the &battunement'
-  - 'of this weapon. Dark magic is cool, right?'
+  - "Shoot a projectile based on the &battunement"
+  - "of this weapon. Dark magic is cool, right?"
 
 ability-name-tarhelm: Tarhelm
 
 ability-lore-tarhelm:
-  - 'Gives you &cstrength&7, but drastically lowers your walk speed.'
+  - "Gives you &cstrength&7, but drastically lowers your walk speed."
 
 ability-name-crucify: Crucify
 
 ablity-lore-crucify:
-  - 'Disables all &cregeneration&o, but gives a'
-  - 'permanent &cstrength &7and &fspeed &7boost.'
+  - "Disables all &cregeneration&o, but gives a"
+  - "permanent &cstrength &7and &fspeed &7boost."
+  - ""
+  - "On death, this item and it's effects will disappear and"
+  - "create a large explosion, destroying everything in its path"
 
 ability-name-centralize: Centralize
+ability-name-centralize_corrupt: Centralize
 
 ability-lore-centralize:
-  - 'Place down a &e&lTotem &7which temporarily'
-  - 'grants &cResistance I &7and &6Fire Resistance I'
-  - 'to all players in a &a30 block &7radius.'
+  - "Place down a &e&lTotem &7which temporarily"
+  - "grants &cResistance I &7and &6Fire Resistance I"
+  - "to all players in a &a30 block &7radius."
     
-ability-lore-centralize-corrupt:
-  - 'Place down a &5&lTotem&7 which temporarily'
-  - 'grants &cResistance II&7 and &6Fire Resistance I'
-  - 'to you, yourself in a &a45 block &7radius.'
+ability-lore-centralize_corrupt:
+  - "Place down a &5&lTotem&7 which temporarily"
+  - "grants &cResistance II&7 and &6Fire Resistance I"
+  - "to you, yourself in a &a45 block &7radius."
 
 ability-name-void_warp: Void Warp
 
 ability-lore-void_warp:
-  - 'Creates a &drift &7in space and pulls you through it.'
+  - "Creates a &drift &7in space and pulls you through it."
 
 ability-name-void_recall: Voidular Recall
 
 ability-lore-void_recall:
-  - 'Binds a nearby block to your &bsoul.'
-  - 'Teleports you back to the &bsoulbound &7block.'
+  - "Binds a nearby block to your &bsoul."
+  - "Teleports you back to the &bsoulbound &7block."
+
+ability-name-obsidian_shard: Obsidian Shards
+
+ability-lore-obsidian_shard:
+  - "Shoots a extremely sharp shard of &8obsidian."
+  - "Deals massive &cdamage &7on impact."
+
+ability-name-justice: Justice
+
+ability-lore-justice:
+  - "Strikes &flighting &7and gives &cRegeneration I&7 on hit."
